fix MSVC warning: make sizeCoefficient double (was float)

+++ app/src/libraries/helpers/ConfigEditorHelper.cpp
@@ -18,7 +18,7 @@ ConfigEditorHelper::ConfigEditorHelper()
 
 
 // Редактирование произвольного конфиг-файла (конфиг программы, конфиг редактора)
-void ConfigEditorHelper::editConfigFile( QString fileName, float sizeCoefficient )
+void ConfigEditorHelper::editConfigFile( QString fileName, double sizeCoefficient )
 {
   // Окно диалога для редактирования файла конфига
   EditorMultiLineInputDialog dialog( qobject_cast<QWidget *>(pMainWindow) );
